EventForm and styling changes
The event form's data is now able to be saved similarly to the event creation regardless if they close the editor or not. The options to add the buttons and questions are all good as well. The Prisma should be up to date and all for a user form as that can be pulled/shown for the user side to be filled in now :) Maybe a little more UI/UX could be done for the next and back buttons. I also think I should create a form submission inside the EventEditor instead of using the publish button as it is becoming a hydration issue with a form within a form. I will also have to push my scheme to the database and try and submit a form. I will also do more work on the eventEditor side and add an edit button and hopefully a edit form button and make the event and the form into one link. I might do a draft button for the event as well. I will also work on the styling of the event cards that are being displayed. I will add the image functionality that is displayed in the Articles cards as well.